requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pynacl>=1.5.0
zeroconf>=0.131.0
httpx>=0.25.0
pydantic>=2.5.0
